Feat/recconect and 2phase

+++ user/app/controllers/file_operations.py
@@ -40,28 +40,24 @@ async def upload_file(file: UploadFile = File(...), name: str = Form(...), path:
             for server in servers:
                 server_chunks[server['url']].append((chunk, chunk_hash))
 
-    # async with aiohttp.ClientSession() as session:
-    for server_url, chunks in server_chunks.items():
-        for i in range(0, len(chunks), MAX_CHUNKS_PER_REQUEST):
-            batch_chunks = chunks[i:i + MAX_CHUNKS_PER_REQUEST]
-            url = f"{server_url}/store_chunks_pending/"
-            files = []
-            chunks = [
-            {"data": chunk.decode('utf-8'), "chunk_hash": chunk_hash} for chunk, chunk_hash in batch_chunks
-            ]
-        response = requests.post(url, json=chunks)
-        if response.status_code != 200:
-            raise Exception(f"Failed to store chunks on server {server_url}, status code: {response.status_code}")
-                # async with session.post(url, data=data) as response:
-                #     if response.status != 200:
-                #         raise Exception(f"Failed to store chunks on server {server_url}, status code: {response.status}")
+    async with aiohttp.ClientSession() as session:
+        for server_url, chunks in server_chunks.items():
+            for i in range(0, len(chunks), MAX_CHUNKS_PER_REQUEST):
+                batch_chunks = chunks[i:i + MAX_CHUNKS_PER_REQUEST]
+                url = f"{server_url}/store_chunks_pending/"
+                data = [
+                    {"data": chunk.decode('utf-8'), "chunk_hash": chunk_hash} for chunk, chunk_hash in batch_chunks
+                ]
+                async with session.post(url, json=data) as response:
+                    if response.status != 200:
+                        raise Exception(f"Failed to store chunks on server {server_url}, status code: {response.status}")
 
 
-    # Notify the leader about the new file and its chunks
+    # Notify the leader about the new file and its chunks as temporary
     name_mapping = {"full_path": os.path.join(path, name), "chunk_hashes": chunk_positions, "size": file_size}
-    response = requests.post(f"{os.getenv('LEADER_URL')}/namemappings/", json=name_mapping)
+    response = requests.post(f"{os.getenv('LEADER_URL')}/temp_namemappings/", json=name_mapping)
     if response.status_code != 200:
-        raise HTTPException(status_code=response.status_code, detail="Error creating name mapping")
+        raise HTTPException(status_code=response.status_code, detail="Error creating temporary name mapping")
 
     # Finalize chunks on the servers
     async with aiohttp.ClientSession() as session:
@@ -72,6 +68,11 @@ async def upload_file(file: UploadFile = File(...), name: str = Form(...), path:
                 async with session.post(url, json={"chunks": [chunk_hash]}) as response:
                     if response.status != 200:
                         raise HTTPException(status_code=response.status, detail=f"Failed to finalize chunk on server {server['url']}")
+
+    # Finalize name mapping on the leader
+    response = requests.post(f"{os.getenv('LEADER_URL')}/finalize_namemappings/", params={"full_path": os.path.join(path, name)})
+    if response.status_code != 200:
+        raise HTTPException(status_code=response.status_code, detail="Error finalizing name mapping")
 
     os.remove(file_location)  # Cleanup the temporary file
     return {"message": "File uploaded and processed successfully"}
